+++ app/upgrades/v1/constants.go
@@ -1,4 +1,4 @@
-package v2
+package v1
 
 const (
 	// UpgradeName is the shared upgrade plan name for mainnet
