+++ app/upgrades/v1/upgrades.go
@@ -1,4 +1,4 @@
-package v2
+package v1
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
